Fix setBufferInt.

+++ support/c/idris_buffer.h
@@ -16,7 +16,7 @@ void idris2_freeBuffer(void* buf);
 int idris2_getBufferSize(void* buffer);
 
 void idris2_setBufferByte(void* buffer, int loc, int byte);
-void idris2_setBufferInt(void* buffer, int loc, int val);
+void idris2_setBufferInt(void* buffer, int loc, int64_t val);
 void idris2_setBufferDouble(void* buffer, int loc, double val);
 void idris2_setBufferString(void* buffer, int loc, char* str);
